add facture

+++ src/main/java/ma/dentaire/projetdentaires8/controller/ConsultationController.java
@@ -24,7 +24,7 @@ public class ConsultationController {
     private IServiceConsultation serviceConsultation;
 
     @GetMapping("/patient/consultation/{id}")
-    public String showAddPatientForm(Model model, @PathVariable Integer id) {
+    public String showAddPatientForm(Model model, @PathVariable Long id) {
         PatientInfoDto patient = servicePatient.findPatientInfos(id);
         List<ConsultationShowDto> consultations = serviceConsultation.findPatientConsultations(id) ;
